+++ lib/src/models/types/data/text.dart
@@ -5,22 +5,25 @@ import 'dart:convert';
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import '../serializer/serializers.dart';
-import 'renderizable.dart';
+import '../../serializer/serializers.dart';
+import '../rich_text/generic_text.dart';
 import 'span.dart';
 
 part 'text.g.dart';
 
 ///Text model
 @BuiltValue(wireName: 'paragraph')
-abstract class Text implements Renderizable, Built<Text, TextBuilder> {
+abstract class Text implements GenericText, Built<Text, TextBuilder> {
   ///Current text
+  @override
   String get text;
 
   ///List of spans
+  @override
   BuiltList<Span> get spans;
 
   ///Type of
+  @override
   @nullable
   String get type;
 
@@ -31,13 +34,12 @@ abstract class Text implements Renderizable, Built<Text, TextBuilder> {
 
   ///Converts to json
   String toJson() {
-    return json
-        .encode(richTextSerializers.serializeWith(Text.serializer, this));
+    return json.encode(flusmicSerializers.serializeWith(Text.serializer, this));
   }
 
   ///Converts from json
   static Text fromJson(String jsonString) {
-    return richTextSerializers.deserializeWith(
+    return flusmicSerializers.deserializeWith(
         Text.serializer, json.decode(jsonString));
   }