+++ lib/src/models/types/info/language.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import '../serializer/serializers.dart';
+import '../../serializer/serializers.dart';
 
 part 'language.g.dart';
 
@@ -23,12 +23,13 @@ abstract class Language implements Built<Language, LanguageBuilder> {
 
   ///Converts to json
   String toJson() {
-    return json.encode(serializers.serializeWith(Language.serializer, this));
+    return json
+        .encode(flusmicSerializers.serializeWith(Language.serializer, this));
   }
 
   ///Converts from json
   static Language fromJson(String jsonString) {
-    return serializers.deserializeWith(
+    return flusmicSerializers.deserializeWith(
         Language.serializer, json.decode(jsonString));
   }