+++ lib/src/models/types/rich_text/generic_image.dart
@@ -1,0 +1,16 @@
+import '../info/dimension.dart';
+
+///Base class for images
+abstract class GenericImage {
+  ///Image dimensions
+  Dimension get dimensions;
+
+  ///Alt
+  String get alt;
+
+  ///Copyright
+  String get copyright;
+
+  ///Image url
+  String get url;
+}